+++ ts/packages/providers/google/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @composio/google
 
+## 0.6.11
+
+### Patch Changes
+
+- Updated dependencies [27ed0c9]
+  - @composio/core@0.6.11
+
 ## 0.6.10
 
 ### Patch Changes